Refactor extraction logic and use relative paths
